Fixed error importing Illinois campaign contributions data
in the response, a method read() was being called that didn't exist on the object.  The correct attribute is content.  It returns binary content, which is what we want for downloading a zip file.  We then at the flag 'b' to the file we open to let it know we will be writing binary data